Make fast library work on Windows
